(PC-40397)[ADAGE] fix: send int instead of float for durationMinutes on offre-assoc call

+++ api/src/pcapi/core/educational/adage_backends/serialize.py
@@ -29,7 +29,7 @@ class AdageCollectiveOffer(schemas.AdageBaseResponseModel):
     endDatetime: datetime
     contact: AdageCollectiveOfferContact
     description: str | None
-    durationMinutes: float | None
+    durationMinutes: int | None
     id: int
     name: str
     numberOfTickets: int
